refactor: moify route structure
