--- backend/routes/about-routes.js
@@ -1,9 +1,0 @@
-import express from "express";
-
-const router = express.Router();
-
-router.get("/", (req, res, next) => {
-  res.json({});
-});
-
-export default router;
